requirements.txt
django==5.1.7
psycopg2==2.9.10
pymongo==3.11.4
celery==5.4.0
django-celery-beat==2.7.0
django-allauth==65.5.0
djangorestframework==3.14.0
requests==2.31.0
django-allauth[socialaccount]
PyJWT
faker


